pyproject.toml
[tool.poetry]
name = "flightrobustness"
version = "0.1.0"
description = "Monte Carlo Simulator for Cascading Flight Delays"
authors = ["Rafique Nazir <rafique.nazir@gmail.com>"]
readme = "README.md"
packages = [{ include = "flightrobustness", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"
polars = "^0.20.0"
matplotlib = "^3.8.0"
pyyaml = "^6.0.1"
fastapi = {extras = ["standard"], version = "^0.118.0"}
uvicorn = { extras = ["standard"], version = "^0.30.0" }
boto3 = "^1.35.0"
python-dotenv = "^1.0.1"
pydantic = "^2.8.0"
aiofiles = "^23.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
simulate-cli = "flightrobustness.interfaces.cli:main"
simulate-api = "flightrobustness.interfaces.api:app"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --tb=short --disable-warnings"
testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
log_cli = true
log_cli_level = "INFO"
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
